project_5140_energy/requirements.txt
pandas>=1.5.0
numpy>=1.24.0
matplotlib>=3.6.0
statsmodels>=0.14.0
scipy>=1.10.0
python-docx>=0.8.11
prophet>=1.1.0
